Lazy load drawer and improve styles
