+++ docs/src/components/MobileNavContext.ts
@@ -1,0 +1,13 @@
+'use client';
+import * as React from 'react';
+import type { Drawer } from '@base-ui/react/drawer';
+
+export const MobileNavContext = React.createContext<{
+  handle: Drawer.Handle<unknown> | null;
+  searchValue: string;
+  setSearchValue: (value: string) => void;
+}>({
+  handle: null,
+  searchValue: '',
+  setSearchValue: () => undefined,
+});
